Mayor changes in utilities
Cifti_utils was splited in cifti_utils and cifti_header, this will
allow to separate the functionality of those scripts (one to manipulate
cifti structures in general, the other to create cifti headers). The
script other_utils was removed. The tests were modified to reflect
the changes in the utilities.