Fix small bug in cifti_filter
Fixed a small bug in cifti filtered. It happened when the indices to
search for where not present in the file to filter. In that case, the
last column was appended instead, since the "filter indices" function
in "cifti utils" was returning -1, we should considere changing
that behaviour to None maybe"