Fix TRK saving format
This commit fixes the streamline_utils.save function in order to
make it (hopefully) fully compatible with the trackvis software.
While in the previous version the streamlines were correctly stored
and loaded, the trackvis software had problems in visualize them.
The problems arises from nibabel, since it doesn't write the correct
header into the trk file.