Make cifti_parcellate compatible with latest packages
I updated the cifti_parcellate script so now it works with the last
version of numpy, nibabel, and it is python 3 compatible. In order to do
so, a dependency to my package "citrix" has been added. This also
allowed me to separate the handling of CIFTI information away from
logpar. Logpar will be only a parcellation package.